refactor(agents): modularize pipeline into types, prompt, transform, and orchestrator
